+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - PBAP Client: fix PBAP_SUBEVENT_OPERATION_COMPLETED with OBEX_DISCONNECTED for pbap_disconnect
 - HFP HF: send HF Indicator update only if enabled by AG
 - HFP AG: always initiate codec connection setup before establishing audio
+- HID Host: omit Report ID in Set/Get Report and send report for report id == HID_REPORT_ID_UNDEFINED
 - POSIX: clear run loop exit flag
  
 ### Changed